미완성 수정본

+++ Video/views.py
@@ -6,6 +6,8 @@ from django.http import JsonResponse
 from django.http import HttpResponseRedirect
 from django.shortcuts import render, redirect
 # from Video.forms import VideoForm
+from django.views.decorators.http import require_POST
+
 from Video.models import Video
 
 from .forms import VideoForm
@@ -25,31 +27,31 @@ def getTag(request, tags):
 
     return JsonResponse({'message':json})
 
-
-
-
-
-
+@login_required(login_url='/users/login')
 def upload(request):
     if request.method == "GET":
         videoform = VideoForm()  # 입력칸에 아무것도 없는 상태로 준다
         return render(request, 'video/upload.html',
                       {'form': videoform})
     elif request.method == 'POST':
-        request.FILES
-
         videoform = VideoForm(request.POST, request.FILES)
+
         if videoform.is_valid():
-            video = videoform.save()
+            video = videoform.save(commit=False)
+            video.writer = request.user
             video.save()
             return redirect('/Video/mylist')
         else:
             videoform = VideoForm()
-    return render(request, 'Video/upload.html', {'form': videoform})
+            return render(request, 'Video/upload.html', {'form': videoform})
+
+
+
+
 
 def ss(request):
 
-    return render(request, 'layout/base.html')
+    return render(request, 'Video/list.html')
 
 def posts(request):
     posts = Video.objects.all()  # board table에서 모든 데이터를 다 가져옴
@@ -70,6 +72,7 @@ def read(request, bid) :
     return render(request, 'Video/read.html',{'read' : post, 'posts':posts })
     # #board/read.html페이지를 보여주고 Board.objects.get( Q(id=bid))의 값
     # 을 'read'로 저장한다.
+
 def readmine(request, bid) :
     post = Video.objects.get( Q(id=bid))
     #posts = Board.objects.all()
@@ -102,30 +105,42 @@ def list_edu(request):
     return render(request, 'video/list_edu.html',
                   {'posts': posts})
 
+
+
+
+@login_required(login_url='/users/login')
 def delete(request, bid) :
     post = Video.objects.get(Q(id = bid))
+    if request.user != post.writer:
+        return render(request,'Video/list.html')
     post.delete()
     return redirect('/Video/list')
 
 
+
+@login_required(login_url='/users/login')
 def update(request, bid) :
     post = Video.objects.get(Q(id=bid))  #게시글 하나를 가져오는것
     if request.method == "GET" :
-        videoForm = VideoForm(instance=post)
+        videoForm = VideoForm()
         #특정 조건 id에 해당하는 값을 저장한다.
         return render(request,'Video/update.html',{'videoForm' : videoForm})
     elif request.method == "POST" :
         videoForm = VideoForm(request.POST, request.FILES)
-
-        # boardForm에서 사용자가 보내온 데이터를 받느다
-        if videoForm.is_valid():   #boardForm안에 값이 유효한다면
+        print(videoForm.cleaned_data['title'])
+        if videoForm.is_valid():   #boardForm안에 값이 유효한다
             post.title = videoForm.cleaned_data['title']
             post.tag = videoForm.cleaned_data['tag']
-            # post.file = videoForm.cleaned_data['file']
+            post.file = videoForm.cleaned_data['file']
+            post.file2 = videoForm.cleaned_data['file2']
+            post.writer = request.user
             post.save()
+            print("1")
             return redirect('/Video/list')
 
             # return render(request, 'Video/list.html',)
+        else:
+            return redirect('/')
